requirements_dev.txt
backports.zoneinfo==0.2.1
bleach==4.1.0
boto3==1.18.35
botocore==1.21.35
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
# Editable install with no version control (cloudgeass==0.1.1)
-e c:\users\thipa\appdata\local\packages\pythonsoftwarefoundation.python.3.8_qbz5n2kfra8p0\localcache\local-packages\python38\site-packages
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.7.1
dnspython==2.1.0
docutils==0.17.1
exchangelib==3.3.0
idna==3.2
importlib-metadata==4.8.1
isodate==0.6.0
jaiminho==0.0.5
jmespath==0.10.0
keyring==23.1.0
lxml==4.6.3
nba-api==1.1.9
-e git+https://github.com/ThiagoPanini/nbaflow.git@f8c91f84f20df70b11947d66e909a88d492bed3e#egg=nbaflow
ntlm-auth==1.5.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
pkginfo==1.7.1
pretty-html-table==0.9.dev0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.26.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
s3transfer==0.5.0
six==1.16.0
tqdm==4.62.2
twine==3.4.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.0
